# server/openapi.yaml
openapi: 3.0.3
info:
  title: Digital Scout Camp API
  description: Backend API for the Digital Scout Camp festival management system.
  version: 1.0.0
  contact:
    name: Digital Scout Camp Team
servers:
  - url: /api/v1
    description: Versioned API
  - url: /api
    description: Backward-compatible alias
paths:
...
  /competitions:
    get:
      summary: List competitions
      tags: [Competitions]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated competitions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /competitions/{idOrSlug}/enter:
    post:
      summary: Enter a competition
      tags: [Competitions]
      security:
        - bearerAuth: []
      parameters:
        - name: idOrSlug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                entryCode: { type: string }
      responses:
        '200':
          description: Competition entered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  total: { type: number }
                  scoreId: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PaginatedResponse:
      type: object
      properties:
        success: { type: boolean }
        data: { type: array, items: { type: object } }
        pagination:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
...